% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_cohort_data.r
\name{af_cohort_data}
\alias{af_cohort_data}
\title{Get Data from AppsFlyer Cohort API}
\usage{
af_cohort_data(
  app_id,
  cohort_type,
  min_cohort_size = 1,
  from,
  to,
  granularity = "day",
  partial_data = T,
  filters = NULL,
  groupings,
  kpis,
  preferred_currency = F,
  preferred_timezone = F,
  aggregation_type,
  per_user = F,
  signature
)
}
\arguments{
\item{app_id}{AppsFlyer App ID}

\item{cohort_type}{Cohort attribution (conversion) type}

\item{min_cohort_size}{Minimum cohort size}

\item{from}{Lower bound of the LTV attribution date range}

\item{to}{Higher bound of the LTV attribution date range}

\item{granularity}{Data granularity}

\item{partial_data}{Partial data}

\item{filters}{Filters}

\item{groupings}{Groupings}

\item{kpis}{Selected KPI}

\item{preferred_currency}{KPI revenue currency}

\item{preferred_timezone}{Time zone of data ranges}

\item{aggregation_type}{Aggregation type}

\item{per_user}{Data per user}

\item{signature}{Signature}
}
\description{
Get Data from AppsFlyer Cohort API <https://support.appsflyer.com/hc/en-us/articles/360004799057-Cohort-API>
}
