\name{rappsflyer-package}
\alias{rappsflyer-package}
\alias{rappsflyer}
\docType{package}
\title{
\packageTitle{rappsflyer}
}
\description{
\packageDescription{rappsflyer}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{rappsflyer}
\packageIndices{rappsflyer}
}
\author{
\packageAuthor{rappsflyer}
Maintainer: \packageMaintainer{rappsflyer}
}
\keyword{ package }

