\name{af_set_api_token}
\alias{af_set_api_token}
\title{
Set AppsFlyer API token
}
\description{
Set AppsFlyer API token in current R session
}
\usage{
af_set_api_token(api_token)
}
\arguments{
  \item{api_token}{
Your AppsFlyer API token.
}
}
\author{
Alexey Seleznev
}
\examples{
\dontrun{
af_set_api_token("Your API token")
}
}
